Revert 613ae39, potential memory leak fixed
HashMap does not store copies of the Player object, but references to it.
However, the garbage collector will not free the object's memory if there is at least one reference to it.
To prevent this we must check if the player is online and if not remove it from the Map.